requirements.txt
aioserial~=1.3.0
git+https://github.com/mch/python-ant.git@02e0458#egg=ant
asyncio-mqtt==0.12.1
bleak~=0.14.2
bleak-sigspec==0.0.4
dbus-next==0.2.3
flexpolyline==0.1.0
gpxpy==1.5.0
msgpack-python==0.5.6
paho-mqtt==1.6.1
pyserial~=3.5
pyusb==1.2.1
six==1.16.0
pip~=20.3.4
setuptools~=44.1.1